+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## What's New?
 
+### v1.0.1  
+1. Change config `chunkHash` to `hashDigestLength`, default 8. Or set 0 to close the hash.  
+2. Add `bundleAnalyze` config, set to true to active visualize size of webpack output files with an interactive zoomable treemap. Default to `false`.
+
 ### v1.0.0  
 Help info complete!!!